recipe-automation/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-wordpress-xmlrpc>=2.3
openai>=1.12.0
beautifulsoup4>=4.12.0
